backend/requirements.txt
# === Core Dependencies ===
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.5
pydantic-settings==2.7.0

# === AWS SDK ===
boto3==1.35.90
botocore==1.35.90

# === Gemini Vision ===
google-genai>=0.2.0
Pillow>=10.0.0
numpy>=1.24.0  # Required for TFLite and image processing

# === Utilities ===
python-multipart==0.0.20  # For file uploads
python-jose[cryptography]==3.3.0  # For JWT authentication
passlib[bcrypt]==1.7.4  # For password hashing
bcrypt<4.0  # passlib 1.7.4 compatibility (bcrypt 4.x removed __about__)
cryptography==44.0.0  # For JWT
pydantic[email]==2.10.5  # Email validation support

# === Production ===
gunicorn==23.0.0  # Production WSGI server
python-dotenv==1.0.1  # Load environment variables

# === Database ===
sqlalchemy==2.0.36
# Use psycopg2-binary for development (for production, use psycopg2)
# Note: If installation fails on Windows, you may need to install PostgreSQL development tools
# or use: pip install psycopg2-binary --only-binary=psycopg2-binary
psycopg2-binary==2.9.10
